Adds BookshelfObject enum
This enum holds the instances of Bookshelfobjects available in the game.
The class makes available a static method to return a random
BookshelfObject form the enum.
To comply with the game rules, the number of instances of every object
must be checked. My Shelfy allows 22 instances of the same object.
If the randomly selected object is not available, the next object is
selected.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>