Adds Board class
This class represent a generic board for the game.
It contains a read-only tiles table that define the shape of the board.
It also contains a method to initialize the board, and a method to
randomly select the shelf-objects and place them in the right tiles
for the selected players number.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>