Adds the classes to implement the MVC design patter
The MVC design pattern uses three different classes of objects to
create a level of abstraction good enough to separate the business logic
from the graphical output, and the same for the data used in the process.
In this implementation the GameManager acts like the model, since it
contains all the data needed for the game.
This commit also adds two dedicated exceptions to help the programmers to
correctly implement the pattern.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>