Implements two different version of MainPageView class
This MainPageView classes are examples of how to implement two different
views for the same page. One of the classes (package gui.view.console) uses
the console to output its information and the Scanner class to take inputs
from the user.
The other class (package gui.view.graphic) uses java.swing and java.awt
objects to build a graphic user interface.
Both class use the same Controller, so the business logic behind the ui is
the same.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>