Modifies the UI related classes after the lesson about the MVC pattern
The Controller class is now the connection point between the Model and
the View objects. Therefore, the View object is now an interface since it
has no data but only methods. A new layer of abstraction has been added
to a View, to reduce the needs of code maintenance and to reduce the
possibility of coding errors. This layer is located in the gui.view package
and is an abstract class containing the listeners needed to mange the user
input. Removed the exception related to the null-controller and added the
exception to avoid passing null views.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>