Changes the visibility of attributes and adds method add() to Matrix class
The rows, cols and items attributes are better managed if private.
The rows (number of row in the matrix) and the cols (number of columns in
the matrix) should not be modified by anyone after the construction of the
matrix. Moreover, the items attribute should not be modified directly since
the row and column coordinates passed to the matrix to store the item,
would not be checked. For this reason the add() method has been
implemented.
(Also fixed formatting of if-else statement)

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>