Modifies InsertPlayersView class and all the related classes
To allow the creation of multiple user interfaces, the InsertPlayersView
has been modified to better abstract its methods.
Consequently, the InsertPlayersController has been modified to call these
methods and to remove any type of link with a specific ui implementation.
Finally the two implementation of the view implement the abstract methods.
Due to this modification, the GameManager class does not need to make
available a method to show its players.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>