Modifies the pickObject method of the Board interface and related classes
The board does not need to know how many objects have been selected by the
user, but knows only if an object can be picked from a tile or not.
For this reason, the check of how many objects the user has selected is
a contract that the ui must fulfill.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>