Fixes bug that cause a crash during the put of the objects in the bookshelf
The bug happen if you try to put too many objects in a column already
filled or partially filled and after that you put them in a column
that can contain them. The bug highlights a problem with the
isThereEnoughSpace method and the check-before-put controls during the put
of the objects.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>