Overloads isThereEnoughSpace method
This new signature of the isThereEnoughSpace method takes only the number
of the objects that the player wants to add to the bookshelf, ignoring
the column. This because the method checks the entire bookshelf to
find if globally there is enough space.
This to avoid the case where the player picks more objects than globally
available in the bookshelf, causing the stall of the game.

Signed-off-by: a.zanetti6 <a.zanetti6@studenti.unibg.it>